==> notebooks/01_Exploracion_corpus.ipynb <==
# Exploracion y limpieza de corpus

import pandas as pd

df = pd.read_csv("C:\\Users\\alvar\\OneDrive\\Escritorio\\TFM\\data\\crudo\\initial_corpus.csv")
df.head(15)

df.sample(10)

# inspeccion general
df.info()

# valores nulos por columna
df.isnull().sum()

# verificar longitud de abstracts
df['abstract'].str.len().describe()

# verificacion de abstracts solos
abstracts = df['abstract'].tolist()
ruta_salida = "C:\\Users\\alvar\\OneDrive\\Escritorio\\TFM\\data\\abstracts.txt"
# Separador que se usará entre abstracts
separador = "\n\n========================================\n\n"

with open(ruta_salida, 'w', encoding='utf-8') as f:
        for i, abstract in enumerate(abstracts):
            # Escribir el abstract actual
            f.write(abstract)
            # Escribir el separador (excepto después del último)
            if i < len(abstracts) - 1:
                f.write(separador)
