==> notebooks/02_Pruebas_preprocesado.ipynb <==
# Prueba preprocesado corpus

import pandas as pd
import numpy as np
import os
import sys

module_path = os.path.abspath(os.path.join('..'))
if module_path not in sys.path:
    sys.path.append(module_path)

from src.procesamiento_corpus import preprocess_text, get_text_source

# carga del data frame
df_corpus = pd.read_csv("C:\\Users\\alvar\\OneDrive\\Escritorio\\TFM\\data\\crudo\\initial_corpus.csv")

print("--- Corpus de Prueba Original ---")
display(df_corpus)

# prueba de funcion preprocess_text
# Esta función se encarga de limpiar el texto, eliminando etiquetas HTML, espacios extra y convirtiendo a minúsculas.
sample_text_html = "This has <B>Bold</B> and <I>italic</I> tags.  Plus extra   spaces."
sample_text_clean = "this is already clean"
sample_text_none = None

print(f"Original: '{sample_text_html}' -> Procesado: '{preprocess_text(sample_text_html)}'")
print(f"Original: '{sample_text_clean}' -> Procesado: '{preprocess_text(sample_text_clean)}'")
print(f"Original: '{sample_text_none}' -> Procesado: '{preprocess_text(sample_text_none)}'")

# prueba get_text_source sobre el corpus de prueba
# Esta función internamente usa `preprocess_text` y la lógica de
# seleccionar entre abstract y title.

min_len_abstract = 50 # Longitud mínima de palabras para considerar un abstract

df_corpus['processed_text'] = df_corpus.apply(
    lambda row: get_text_source(row, min_abstract_len=min_len_abstract),
    axis=1
)

print(f"\n--- Corpus de Prueba con Columna 'processed_text' (min_abstract_len={min_len_abstract}) ---")
display(df_corpus[['paperId', 'title', 'abstract', 'processed_text']])

# verificacion de abstracts procesados solos
abstracts = df_corpus['processed_text'].tolist()
ruta_salida = "C:\\Users\\alvar\\OneDrive\\Escritorio\\TFM\\data\\procesado\\abstracts_procesados.txt"
# Separador que se usará entre abstracts
separador = "\n\n========================================\n\n"

with open(ruta_salida, 'w', encoding='utf-8') as f:
        for i, abstract in enumerate(abstracts):
            # Escribir el abstract actual
            f.write(abstract)
            # Escribir el separador (excepto después del último)
            if i < len(abstracts) - 1:
                f.write(separador)
