==> notebooks/03_generacion_embeddings.ipynb <==
# Generacion de embeddings

# configuracion y importaciones
import pandas as pd
import numpy as np
import os
import pickle # Para guardar la lista de IDs
import sys
import re
import ast

# Ajuste del sys.path para importar módulos de src 
module_path = os.path.abspath(os.path.join('..'))
if module_path not in sys.path:
    sys.path.append(module_path)

# Importar la función para generar embeddings
from src.generador_embeddings import generate_sbert_embeddings

# Procesado del corpus previo generacion de embeddings
from src.procesamiento_corpus import get_text_source
df_corpus = pd.read_csv("C:\\Users\\alvar\\OneDrive\\Escritorio\\TFM\\data\\crudo\\initial_corpus.csv")
min_len_abstract = 50 # Longitud mínima de palabras para considerar un abstract sino se toma el titulo del articulo

df_corpus['processed_text'] = df_corpus.apply(
    lambda row: get_text_source(row, min_abstract_len=min_len_abstract),
    axis=1)



# limpieza del campo openAccesPdf
from src.procesamiento_corpus import extract_pdf_url_from_dict_string

df_corpus['pdf_url'] = df_corpus['openAccessPdf'].apply(extract_pdf_url_from_dict_string)
df_corpus


# guardo el corpus procesado
df_corpus.to_csv("C:\\Users\\alvar\\OneDrive\\Escritorio\\TFM\\data\\Procesado\corpus_procesado.csv", index=False, encoding='utf-8')

# Preparar la lista de textos y la lista de IDs (en el mismo orden)
texts_to_encode = df_corpus['processed_text'].tolist()
paper_ids_ordered = df_corpus['paperId'].tolist()

print(f"Número de textos a codificar: {len(texts_to_encode)}")
print(f"Número de IDs correspondientes: {len(paper_ids_ordered)}")

# generar los embeddings con SBERT

# definir modelo a utilizar:
# 'all-MiniLM-L6-v2' es bueno y rapido.
# Otros podrían ser: 'multi-qa-mpnet-base-dot-v1' (bueno para búsqueda semántica)
# o 'allenai-specter' (específico para papers científicos, pero más pesado)
sbert_model_name = 'all-MiniLM-L6-v2'

corpus_embeddings = generate_sbert_embeddings(
        texts_to_encode,
        model_name=sbert_model_name,
        batch_size=32 
    )

# Verificar la forma de los embeddings
# Debería ser (numero de textos, dimension del embedding)
# Para 'all-MiniLM-L6-v2', la dimensión es 384.
print(f"\nForma del array de embeddings: {corpus_embeddings.shape}")
if corpus_embeddings.shape[0] == len(texts_to_encode):
    print("El número de embeddings coincide con el número de textos.")
else:
    print("Error: El número de embeddings NO coincide con el número de textos.")

print(f"Dimensión del embedding para '{sbert_model_name}': {corpus_embeddings.shape[1]}")

# guardar embeddings y los ids
# IMP!! guardar los embeddings en el mismo orden que los ids para despues poder buscarlos correctamente

embeddings_file_path = os.path.join("C:\\Users\\alvar\\OneDrive\\Escritorio\\TFM\\data\\Embeddings", f'corpus_embeddings_{sbert_model_name.replace("/", "_")}.npy')
ids_file_path = os.path.join("C:\\Users\\alvar\\OneDrive\\Escritorio\\TFM\\data\\Embeddings", f'corpus_paper_ids_{sbert_model_name.replace("/", "_")}.pkl')

# Guardar embeddings (array NumPy)
np.save(embeddings_file_path, corpus_embeddings)
print(f"\nEmbeddings guardados en: {embeddings_file_path}")

# Guardar lista de paperIds 
with open(ids_file_path, 'wb') as f:
    pickle.dump(paper_ids_ordered, f)
print(f"Lista de Paper IDs ordenada guardada en: {ids_file_path}")
