==> notebooks/04_pruebas_motor_recomendacion.ipynb <==
# Pruebas motor de recomendacion

import sys
import os
import pandas as pd

# importo las funciones necesarias
module_path = os.path.abspath(os.path.join('..'))
sys.path.append(module_path)
from src.motor_de_recomendaciones import get_recommendations


# Configurar Pandas para mostrar más texto en los outputs
pd.set_option('display.max_colwidth', 200) # Muestra más caracteres de las columnas de texto

# Definir consultas de prueba
queries_de_prueba = [
    "machine learning techniques for natural language processing",
    "impact of social media on society",
    "bibliographic resource recommenders using NLP", 
    "renewable energy sources and climate change"]


for i, query in enumerate(queries_de_prueba):
        print(f"\n======================================================================")
        print(f"EJECUTANDO CONSULTA DE PRUEBA {i+1}/{len(queries_de_prueba)}: '{query}'")
        print(f"======================================================================")
        
        recommendations = get_recommendations(user_query=query, top_n=3) # 3 recomendaciones por consulta

        print(f"\n--- {len(recommendations)} Recomendaciones Encontradas ---")
        for rec in recommendations:
            print(f"  Rank: {rec.get('rank')}")
            print(f"  Paper ID: {rec.get('paperId')}")
            print(f"  Título: {rec.get('title')}")
            print(f"  Año: {rec.get('year')}")
            print(f"  Score de Similitud: {rec.get('similarity_score'):.4f}")
            print(f"  URL PDF: {rec.get('pdf_url', 'No disponible')}")
            print(f"  Abstract: {rec.get('abstract', 'N/A')}")
            print("  -----------------------------------")
